+++ LabelStoreMax/lib/app/providers/example_pay.dart
@@ -53,8 +53,7 @@ examplePay(context,
     showToastNotification(
       context,
       title: trans(context, "Error"),
-      description: trans(context,
-          trans(context, "Something went wrong, please contact our store")),
+      description: trans(context, "Something went wrong, please contact our store"),
     );
     state.reloadState(showLoader: false);
   }