+++ LabelStoreMax/lib/app/providers/stripe_pay.dart
@@ -91,10 +91,8 @@ stripePay(context,
             showToastNotification(
               context,
               title: trans(context, "Error"),
-              description: trans(
-                  context,
-                  trans(context,
-                      "Something went wrong, please contact our store")),
+              description: trans(context,
+                  "Something went wrong, please contact our store"),
             );
             state.reloadState(showLoader: false);
           }